core: auto-generate all code at build time
This moves the core codebase to generate all code related to
FoundationDB at compile time, rather than build time. This gives
more flexible control for clients who want to use specific server
versions.

Notably, foundationdb-sys now allows the user to control the
FDB_API_VERSION setting when calling bindgen. This is exposed
as a set of Cargo features which end-users can pick out: fdb-5_1,
fdb-5_2, and fdb-6_0. This means that foundationdb-sys no longer
has to ship the wrapper.h file, nor a pre-computed bindings.rs file.

Next, it also makes foundationdb-gen a build-time dependency of the
main foundationdb library, and it generates all of the options.rs
module at compile time -- this is because it must be generated based on
fdb.options, which is luckily shipped with all client libraries. Doing
this required moving the generator into a library.

foundationdb-gen was also tweaked slightly to support FDB 5.2, which
introduced a few new special-cases in the option generation code
(Notably, having to rewrite 'AppendIfFit' to the proper name,
'AppendIfFits').

With this in place, I can compile this library on 64-bit Windows for
FoundationDB 5.2 with no modifications, automatically.

Signed-off-by: Austin Seipp <aseipp@pobox.com>